Shift 'bufferPtr' immediately after reading from file
It's much easier to keep track of where the pointer should be if we advance it immediately after "using" its position rather than advance it later on before the next read. We waste some instructions advancing the pointer in some circumstances when a read will never follow but I think it is worth it for the consistency.