Add support for value tokens to JSON lexical analysis
Identifying value tokens required the the ability to peek at the next character in the file stream. I spent some time setting up a char buffer to read the file into before I realized we can accomplish this using a call to 'ungetc()' instead. I'm sure we'll be using buffers for some of the performance adjustments we'll be doing later in the course.