[TEMP] EOF footer for binary answers file
Commented code for placing a footer at the end of the answers file. I originally planned on checking during the loop if we had prematurely reached the end of file and detect whether the answers file and JSON file were out of sync, but there is some other interaction going on that prevents this from being necessary. Archiving the code here as maybe I'll come back to it at some point.