Move base layer files into a common src folder
I think its OK to move these files into a common sources folder as it will eliminate a whole class of errors where I inadvertently edit the wrong copy of the file. Additionally, the base layer is supposed to be extremely general and useful across code-bases so having to iterate on the same instance until it works well for multiple code-bases will be helpful.