Add the ability to toggle timings with a #define
By playing around with the timings, adding and removing categories that don't time anything, it's clear that the "unaccounted" timings category is mostly the result of calling the actual timings themselves. This commit adds a quick way to toggle timings so the total runtime can be compared when timings aren't active.