Experiment with an intermediate struct to hold timings information
Experiment with creating an intermediate struct in-scope that then updates the timings array when the timing is finished. Curious to see whether it makes life easier or whether it affects performance or less than existing solution. Additionally, adopts some of Casey's naming because I like it better and reserves the 0 index in Timings array to represent "no timer".