Add support for nested timings
This unfortunately requires the re-introduction of an intermediate struct and complicates the timings management macros slightly. More elaborate asserts that can be used directly when START_TIMING and END_TIMING are invoked unfortunately break when recursion is involved and have been stripped out. This has the silver-lining of simplifying the timings management macros a fair bit from what they would otherwise need to be.