Revert "Experiment converting 'haversine_token' to use length-based strings"
This reverts commit f939e2bcdb36396eb9369f1c294b480c1b16313b.

The idea behind this experiment was to create a bucket to hold the 'haversine_tokens' and perform the lexing, parsing and haversine distance calculations in separate phases. The problem being there is no good way to map the raw JSON to the number of tokens it will result in, and the 'haversine_token' struct is many times larger than many of the tokens (which are 1 character large). A more fruitful approach will be to keep lexing / parsing in one phase using CStrings and do the haversine distance calculation in a separate phase as it should be doable to roughly estimate how many pairs we have (and setup an arena to hold them).